Intial commit
* audisp plugins: first works on a path file obfuscator

This commit includes a set of functions for declaring, free,
update an obfuscation tree and retrieve its data.

It also includes a command line test program.

the configuration describes the obfuscation at each folder/file part of
the path in order to build a string making the initial data to be
identified only by those with knowledge of the description.